requirements.txt
# local package
-e .

# external requirements
click~=8.1.3
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

zstandard~=0.20.0
setuptools~=65.5.1
beautifulsoup4~=4.11.2
Markdown~=3.4.3
pandas~=1.5.3
seaborn~=0.12.2
spacy~=2.3.9
tqdm~=4.65.0
src~=0.1.0